+++ src/scripts/tantalos.js
@@ -1,4 +1,4 @@
-import { getBiggerDim, getSmallerDim } from '@/renderengine/helper'
+import { getBiggerDim, getSmallerDim } from '@/renderengine/helper/helper'
 
 function tantalos() {
   const water = [36, 44, 53]
